Fixed build warning on Android

+++ src/audio/SDL_audiocvt.c
@@ -292,14 +292,14 @@ static int SDL_ResampleAudio(const int chans, const int inrate, const int outrat
     const int framelen = chans * (int)sizeof(float);
     const int inframes = inbuflen / framelen;
     /* outbuflen isn't total to write, it's total available. */
-    const int wantedoutframes = ((Sint64)inframes) * outrate / inrate;
+    const int wantedoutframes = (int)((Sint64)inframes * outrate / inrate);
     const int maxoutframes = outbuflen / framelen;
     const int outframes = SDL_min(wantedoutframes, maxoutframes);
     float *dst = outbuf;
     int i, j, chan;
 
     for (i = 0; i < outframes; i++) {
-        const int srcindex = ((Sint64)i) * inrate / outrate;
+        const int srcindex = (int)((Sint64)i * inrate / outrate);
         /* Calculating the following way avoids subtraction or modulo of large
          * floats which have low result precision.
          *   interpolation1
